refactor: 화면 이동 기능 추가 및 로그 개선, 폴더 구조 정리, CommonEvent/Effect 제거, mobileMain 분리
