+++ composeApp/src/commonMain/kotlin/com/example/yookcalc/base/BaseViewModel.kt
@@ -1,0 +1,32 @@
+package com.example.yookcalc.base
+
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.launch
+
+abstract class BaseViewModel<S : UiState, E : UiEvent, F : UiEffect>(
+    initialState: S
+) {
+    private val _state = MutableStateFlow(initialState)
+    val state: StateFlow<S> = _state.asStateFlow()
+
+    private val _effect = MutableSharedFlow<F>()
+    val effect: SharedFlow<F> = _effect
+
+    protected val viewModelScope = CoroutineScope(Dispatchers.Main)
+
+    abstract fun onEvent(event: E)
+
+    protected fun setState(reducer: S.() -> S) {
+        _state.value = _state.value.reducer()
+    }
+
+    protected fun sendEffect(effect: F) {
+        viewModelScope.launch { _effect.emit(effect) }
+    }
+}
