+++ composeApp/src/commonMain/kotlin/com/example/yookcalc/base/UiEffect.kt
@@ -1,0 +1,6 @@
+package com.example.yookcalc.base
+
+interface UiEffect
+
+// 화면에 한 번만 실행돼야 하는 일회성 액션 (Toast, SnackBar)
+// 각 화면마다 sealed interface로 정의
