+++ composeApp/src/commonMain/kotlin/com/example/yookcalc/base/UiState.kt
@@ -1,0 +1,6 @@
+package com.example.yookcalc.base
+
+interface UiState
+
+// 화면에 현재 상태(화면에 표시되는 모든 값)를 담는 데이터용 인터페이스
+// 각 화면마다 XXXUiState라는 data class를 만들고 이 인터페이스를 상속받으면 됨
