+++ composeApp/src/commonMain/kotlin/com/example/yookcalc/data/datasource/SampleDataSource.kt
@@ -1,0 +1,4 @@
+package com.example.yookcalc.data.datasource
+
+class SampleDataSource {
+}
