+++ composeApp/src/commonMain/kotlin/com/example/yookcalc/data/repository/SampleRepository.kt
@@ -1,0 +1,5 @@
+package com.example.yookcalc.data.repository
+
+interface SampleRepository {
+    suspend fun getSample(): String
+}
