+++ composeApp/src/commonMain/kotlin/com/example/yookcalc/data/repository/SampleRepositoryImpl.kt
@@ -1,0 +1,5 @@
+package com.example.yookcalc.data.repository
+
+class SampleRepositoryImpl : SampleRepository {
+    override suspend fun getSample(): String = "KMP Sample Data"
+}
