--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/base/UiEffect.kt
@@ -1,6 +1,0 @@
-package com.example.yookcalc.presentation.base
-
-interface UiEffect
-
-// 화면에 한 번만 실행돼야 하는 일회성 액션 (Toast, SnackBar)
-// 각 화면마다 sealed interface로 정의
