--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/base/UiEvent.kt
@@ -1,6 +1,0 @@
-package com.example.yookcalc.presentation.base
-
-interface UiEvent
-
-// 사용자가 화면에서 발생시킨 액션 이벤트를 표현 (버튼클릭, 텍스트 입력 변경 등)
-// 각 화면마다 sealed interface로 정의
