--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/navigation/NavigationActions.kt
@@ -1,25 +1,0 @@
-package com.example.yookcalc.presentation.navigation
-
-import cafe.adriel.voyager.navigator.Navigator
-
-// 필요한 이동 함수들을 계속 추가 가능
-object NavigationActions {
-
-//    Screen -> 이동 함수 추가
-//    object NavigationActions {
-//        fun navigateToSettingScreen(navigator: Navigator) {
-//            navigator.push(SettingScreen)
-//        }
-//    }
-
-//    UI코드에서 호출
-//    @Composable
-//    fun MyPageScreenUI(navigator: Navigator) {
-//        Column {
-//            Button(onClick = { NavigationActions.navigateToSettingScreen(navigator) }) {
-//                Text("설정")
-//            }
-//        }
-//    }
-
-}
