--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/navigation/appTabList.kt
@@ -1,8 +1,0 @@
-package com.example.yookcalc.presentation.navigation
-
-val appTabsList = listOf(
-    AppTab.Maternity,
-    AppTab.Parental,
-    AppTab.ShortenedWork,
-    AppTab.MyPage
-)
