--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/screen/Screen.kt
@@ -1,16 +1,0 @@
-package com.example.yookcalc.presentation.screen
-
-// 나중에 추가할 Voyager Screen 객체들을 위한 템플릿 파일
-// 여기에 새로운 화면(Screen)들을 object로 추가해가면 됨
-
-import androidx.compose.runtime.Composable
-import cafe.adriel.voyager.core.screen.Screen
-
-// 예시 기본 스크린 (삭제해도 무방)
-object SettingScreen : Screen {
-    @Composable
-    override fun Content() {
-        // TODO: SettingScreenUI()
-    }
-    // NavigationActions.kt에서 이동 함수 추가
-}
