--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/screen/mypage/MyPageUI.kt
@@ -1,9 +1,0 @@
-package com.example.yookcalc.presentation.screen.mypage
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun MyPageUI() {
-    Text("마이페이지")
-}
