--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/screen/parental/ParentalUI.kt
@@ -1,9 +1,0 @@
-package com.example.yookcalc.presentation.screen.parental
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ParentalUI() {
-    Text("육아휴직 화면")
-}
