--- composeApp/src/commonMain/kotlin/com/example/yookcalc/presentation/screen/shortened_work/ShortenedWorkUI.kt
@@ -1,9 +1,0 @@
-package com.example.yookcalc.presentation.screen.shortened_work
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ShortenedWorkUI() {
-    Text("육아기 근로시간 단축 화면")
-}
