+++ composeApp/src/iosMain/kotlin/com/example/yookcalc/MainViewController.kt
@@ -2,4 +2,6 @@ package com.example.yookcalc
 
 import androidx.compose.ui.window.ComposeUIViewController
 
-fun MainViewController() = ComposeUIViewController { App() }
+fun MainViewController() = ComposeUIViewController {
+    AppEntry()
+}
