+++ composeApp/src/mobileMain/kotlin/com/example/yookcalc/navigation/appTabList.kt
@@ -1,0 +1,9 @@
+package com.example.yookcalc.navigation
+
+// 실제 탭 목록
+val appTabsList = listOf(
+    AppTab.Maternity,
+    AppTab.Parental,
+    AppTab.ShortenedWork,
+    AppTab.MyPage
+)
