+++ composeApp/src/wasmJsMain/kotlin/com/example/yookcalc/main.kt
@@ -1,5 +1,6 @@
 package com.example.yookcalc
 
+import androidx.compose.material3.Text
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import kotlinx.browser.document
@@ -7,6 +8,6 @@ import kotlinx.browser.document
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
     ComposeViewport(document.body!!) {
-        App()
+        Text("Hello World")
     }
 }
